+++ src/Amadevus.RecordGenerator.Generators/Names.cs
@@ -3,17 +3,20 @@
     internal static class Names
     {
         public const string WithPrefix = "With";
-        
+
         public const string NotSupportedExceptionFull = "System.NotSupportedException";
         public const string Update = "Update";
         public const string UpdateWith = "UpdateWith";
         public const string Deconstruct = "Deconstruct";
-        
+
         public const string Builder = "Builder";
 
         public const string ToBuilder = "ToBuilder";
         public const string ToImmutable = "ToImmutable";
         public const string Validate = "Validate";
+        public new const string ToString = nameof(object.ToString);
+
+        public const string Value = "value";
 
         public const string ToolName = "Amadevus.RecordGenerator";
         public const string GeneratedCodeAttribute = "System.CodeDom.Compiler.GeneratedCodeAttribute";