+++ src/Amadevus.RecordGenerator.Generators/RecordPartialGenerator.cs
@@ -30,6 +30,7 @@ namespace Amadevus.RecordGenerator.Generators
                     GenerateUpdateMethod())
                 .AddRange(
                     GenerateMutators())
+                .Add(GenerateToString())
                 .Add(GenerateValidatePartialMethod());
         }
 
@@ -53,7 +54,7 @@ namespace Amadevus.RecordGenerator.Generators
                                 SyntaxKind.SimpleMemberAccessExpression,
                                 ThisExpression(),
                                 IdentifierName(entry.Identifier)),
-                            IdentifierName(entry.Identifier)));
+                            IdentifierName(entry.IdentifierInCamelCase)));
             }
             StatementSyntax CreateValidateInvocation()
             {
@@ -68,7 +69,7 @@ namespace Amadevus.RecordGenerator.Generators
             {
                 return
                     Argument(
-                        IdentifierName(entry.Identifier))
+                        IdentifierName(entry.IdentifierInCamelCase))
                     .WithRefKindKeyword(Token(SyntaxKind.RefKeyword));
             }
         }
@@ -78,7 +79,7 @@ namespace Amadevus.RecordGenerator.Generators
             var arguments = Descriptor.Entries.Select(x =>
             {
                 return Argument(
-                    IdentifierName(x.Identifier));
+                    IdentifierName(x.IdentifierInCamelCase));
             });
             return MethodDeclaration(Descriptor.Type, Names.Update)
                 .AddModifiers(SyntaxKind.PublicKeyword)
@@ -98,11 +99,14 @@ namespace Amadevus.RecordGenerator.Generators
             return Descriptor.Entries.Select(CreateRecordMutator);
             MethodDeclarationSyntax CreateRecordMutator(RecordDescriptor.Entry entry)
             {
+                var valueIdentifier = Identifier(Names.Value);
+
                 var arguments = Descriptor.Entries.Select(x =>
                 {
                     return Argument(
-                        IdentifierName(x.Identifier));
+                        IdentifierName(x == entry ? valueIdentifier : x.Identifier));
                 });
+
                 var mutator =
                     MethodDeclaration(
                         Descriptor.Type,
@@ -110,7 +114,7 @@ namespace Amadevus.RecordGenerator.Generators
                     .AddModifiers(SyntaxKind.PublicKeyword)
                     .WithParameters(
                         Parameter(
-                            entry.Identifier)
+                            valueIdentifier)
                         .WithType(entry.Type))
                     .WithBodyStatements(
                         ReturnStatement(
@@ -127,6 +131,24 @@ namespace Amadevus.RecordGenerator.Generators
             }
         }
 
+        private MemberDeclarationSyntax GenerateToString()
+        {
+            var properties =
+                from e in Descriptor.Entries
+                select AnonymousObjectMemberDeclarator(IdentifierName(e.Identifier));
+            return
+                MethodDeclaration(
+                    PredefinedType(Token(SyntaxKind.StringKeyword)),
+                    Names.ToString)
+                .AddModifiers(SyntaxKind.PublicKeyword, SyntaxKind.OverrideKeyword)
+                .WithExpressionBody(
+                    InvocationExpression(
+                        MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
+                            AnonymousObjectCreationExpression()
+                                .AddInitializers(properties.ToArray()),
+                            IdentifierName(Names.ToString))));
+        }
+
         private MemberDeclarationSyntax GenerateValidatePartialMethod()
         {
             return
@@ -139,8 +161,8 @@ namespace Amadevus.RecordGenerator.Generators
                 .WithSemicolonToken();
             ParameterSyntax CreateValidateParameter(RecordDescriptor.Entry entry)
             {
-                return 
-                    Parameter(entry.Identifier)
+                return
+                    Parameter(entry.IdentifierInCamelCase)
                     .WithType(entry.Type)
                     .AddModifiers(Token(SyntaxKind.RefKeyword));
             }
@@ -149,7 +171,7 @@ namespace Amadevus.RecordGenerator.Generators
         private static ParameterSyntax CreateParameter(RecordDescriptor.Entry property)
         {
             return Parameter(
-                    property.Identifier)
+                    property.IdentifierInCamelCase)
                 .WithType(property.Type);
         }